dashboard/backend/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
google-auth
pandas
pyarrow
db-dtypes
google-cloud-bigquery-storage